Refactor websocket connection
