+++ packages/next/client/components/react-dev-overlay/hot-reloader-client.tsx
@@ -26,48 +26,11 @@ import {
 } from './internal/error-overlay-reducer'
 import { parseStack } from './internal/helpers/parseStack'
 import ReactDevOverlay from './internal/ReactDevOverlay'
+import { getSocketProtocol } from './internal/helpers/get-socket-protocol'
+import { useErrorHandler } from './internal/helpers/use-error-handler'
+import { useWebsocket } from './internal/helpers/use-websocket'
 
 type DispatchFn = Dispatch<ReducerAction<typeof errorOverlayReducer>>
-
-function onUnhandledError(dispatch: DispatchFn, ev: ErrorEvent) {
-  const error = ev?.error
-  if (!error || !(error instanceof Error) || typeof error.stack !== 'string') {
-    // A non-error was thrown, we don't have anything to show. :-(
-    return
-  }
-
-  if (
-    error.message.match(/(hydration|content does not match|did not match)/i)
-  ) {
-    error.message += `\n\nSee more info here: https://nextjs.org/docs/messages/react-hydration-error`
-  }
-
-  const e = error
-  dispatch({
-    type: ACTION_UNHANDLED_ERROR,
-    reason: error,
-    frames: parseStack(e.stack!),
-  })
-}
-
-function onUnhandledRejection(dispatch: DispatchFn, ev: PromiseRejectionEvent) {
-  const reason = ev?.reason
-  if (
-    !reason ||
-    !(reason instanceof Error) ||
-    typeof reason.stack !== 'string'
-  ) {
-    // A non-error was thrown, we don't have anything to show. :-(
-    return
-  }
-
-  const e = reason
-  dispatch({
-    type: ACTION_UNHANDLED_REJECTION,
-    reason: reason,
-    frames: parseStack(e.stack!),
-  })
-}
 
 function onBuildOk(dispatch: DispatchFn) {
   dispatch({ type: ACTION_BUILD_OK })
@@ -79,17 +42,6 @@ function onBuildError(dispatch: DispatchFn, message: string) {
 
 function onRefresh(dispatch: DispatchFn) {
   dispatch({ type: ACTION_REFRESH })
-}
-
-function getSocketProtocol(assetPrefix: string): string {
-  let protocol = window.location.protocol
-
-  try {
-    // assetPrefix is a url
-    protocol = new URL(assetPrefix).protocol
-  } catch (_) {}
-
-  return protocol === 'http:' ? 'ws' : 'wss'
 }
 
 // const TIMEOUT = 5000
@@ -187,14 +139,6 @@ function tryApplyUpdates(
   sendMessage: any,
   dispatch: DispatchFn
 ) {
-  // @ts-expect-error module.hot exists
-  if (!module.hot) {
-    // HotModuleReplacementPlugin is not in Webpack configuration.
-    console.error('HotModuleReplacementPlugin is not in Webpack configuration.')
-    // window.location.reload();
-    return
-  }
-
   if (!isUpdateAvailable() || !canApplyUpdates()) {
     onBuildOk(dispatch)
     return
@@ -481,7 +425,6 @@ export default function HotReload({
   assetPrefix,
   children,
   initialState,
-  initialTree,
 }: {
   assetPrefix: string
   children?: ReactNode
@@ -491,7 +434,6 @@ export default function HotReload({
   if (initialState?.rootLayoutMissingTagsError) {
     hadRootlayoutError = true
   }
-  const stacktraceLimitRef = useRef<undefined | number>()
   const [state, dispatch] = React.useReducer(errorOverlayReducer, {
     nextId: 1,
     buildError: null,
@@ -499,81 +441,83 @@ export default function HotReload({
     ...initialState,
   })
 
-  const handleOnUnhandledError = useCallback((ev) => {
-    if (
-      ev.error &&
-      ev.error.digest &&
-      (ev.error.digest.startsWith('NEXT_REDIRECT') ||
-        ev.error.digest === 'NEXT_NOT_FOUND')
-    ) {
-      ev.preventDefault()
-      return
-    }
-
-    hadRuntimeError = true
-    onUnhandledError(dispatch, ev)
-  }, [])
-  const handleOnUnhandledRejection = useCallback((ev) => {
-    hadRuntimeError = true
-    onUnhandledRejection(dispatch, ev)
-  }, [])
-
-  const { tree } = useContext(GlobalLayoutRouterContext) ?? {
-    tree: initialTree,
-  }
-  const router = useRouter()
-
-  const webSocketRef = useRef<WebSocket>()
-  const sendMessage = useCallback((data) => {
-    const socket = webSocketRef.current
-    if (!socket || socket.readyState !== socket.OPEN) return
-    return socket.send(data)
-  }, [])
-
-  useEffect(() => {
-    try {
-      const limit = Error.stackTraceLimit
-      Error.stackTraceLimit = 50
-      stacktraceLimitRef.current = limit
-    } catch {}
-
-    window.addEventListener('error', handleOnUnhandledError)
-    window.addEventListener('unhandledrejection', handleOnUnhandledRejection)
-    return () => {
-      if (stacktraceLimitRef.current !== undefined) {
-        try {
-          Error.stackTraceLimit = stacktraceLimitRef.current
-        } catch {}
-        stacktraceLimitRef.current = undefined
+  const handleOnUnhandledError = useCallback(
+    (ev: WindowEventMap['error']): void => {
+      if (
+        ev.error &&
+        ev.error.digest &&
+        (ev.error.digest.startsWith('NEXT_REDIRECT') ||
+          ev.error.digest === 'NEXT_NOT_FOUND')
+      ) {
+        ev.preventDefault()
+        return
       }
 
-      window.removeEventListener('error', handleOnUnhandledError)
-      window.removeEventListener(
-        'unhandledrejection',
-        handleOnUnhandledRejection
-      )
-    }
-  }, [handleOnUnhandledError, handleOnUnhandledRejection])
+      hadRuntimeError = true
+      const error = ev?.error
+      if (
+        !error ||
+        !(error instanceof Error) ||
+        typeof error.stack !== 'string'
+      ) {
+        // A non-error was thrown, we don't have anything to show. :-(
+        return
+      }
 
-  useEffect(() => {
-    if (webSocketRef.current) {
-      return
-    }
+      if (
+        error.message.match(/(hydration|content does not match|did not match)/i)
+      ) {
+        error.message += `\n\nSee more info here: https://nextjs.org/docs/messages/react-hydration-error`
+      }
 
-    const { hostname, port } = window.location
-    const protocol = getSocketProtocol(assetPrefix)
-    const normalizedAssetPrefix = assetPrefix.replace(/^\/+/, '')
+      const e = error
+      dispatch({
+        type: ACTION_UNHANDLED_ERROR,
+        reason: error,
+        frames: parseStack(e.stack!),
+      })
+    },
+    []
+  )
+  const handleOnUnhandledRejection = useCallback(
+    (ev: WindowEventMap['unhandledrejection']): void => {
+      hadRuntimeError = true
+      const reason = ev?.reason
+      if (
+        !reason ||
+        !(reason instanceof Error) ||
+        typeof reason.stack !== 'string'
+      ) {
+        // A non-error was thrown, we don't have anything to show. :-(
+        return
+      }
 
-    let url = `${protocol}://${hostname}:${port}${
-      normalizedAssetPrefix ? `/${normalizedAssetPrefix}` : ''
-    }`
+      const e = reason
+      dispatch({
+        type: ACTION_UNHANDLED_REJECTION,
+        reason: reason,
+        frames: parseStack(e.stack!),
+      })
+    },
+    []
+  )
+  useErrorHandler(handleOnUnhandledError, handleOnUnhandledRejection)
 
-    if (normalizedAssetPrefix.startsWith('http')) {
-      url = `${protocol}://${normalizedAssetPrefix.split('://')[1]}`
-    }
+  const { tree } = useContext(GlobalLayoutRouterContext)
+  const router = useRouter()
+  const webSocketRef = useWebsocket(assetPrefix)
 
-    webSocketRef.current = new window.WebSocket(`${url}/_next/webpack-hmr`)
-  }, [assetPrefix])
+  const sendMessage = useCallback(
+    (data) => {
+      const socket = webSocketRef.current
+      if (!socket || socket.readyState !== socket.OPEN) {
+        return
+      }
+      return socket.send(data)
+    },
+    [webSocketRef]
+  )
+
   useEffect(() => {
     // Taken from on-demand-entries-client.js
     // TODO-APP: check 404 case
@@ -606,26 +550,13 @@ export default function HotReload({
       }
     }
 
-    if (webSocketRef.current) {
-      webSocketRef.current.addEventListener('message', handler)
+    const websocket = webSocketRef.current
+    if (websocket) {
+      websocket.addEventListener('message', handler)
     }
 
-    return () =>
-      webSocketRef.current &&
-      webSocketRef.current.removeEventListener('message', handler)
-  }, [sendMessage, router])
-  // useEffect(() => {
-  //   const interval = setInterval(function () {
-  //     if (
-  //       lastActivityRef.current &&
-  //       Date.now() - lastActivityRef.current > TIMEOUT
-  //     ) {
-  //       handleDisconnect()
-  //     }
-  //   }, 2500)
-
-  //   return () => clearInterval(interval)
-  // })
+    return () => websocket && websocket.removeEventListener('message', handler)
+  }, [sendMessage, router, webSocketRef])
 
   return <ReactDevOverlay state={state}>{children}</ReactDevOverlay>
 }
