+++ packages/next/client/components/react-dev-overlay/internal/helpers/get-socket-protocol.ts
@@ -1,0 +1,10 @@
+export function getSocketProtocol(assetPrefix: string): string {
+  let protocol = window.location.protocol
+
+  try {
+    // assetPrefix is a url
+    protocol = new URL(assetPrefix).protocol
+  } catch (_) {}
+
+  return protocol === 'http:' ? 'ws' : 'wss'
+}
