+++ packages/next/client/components/react-dev-overlay/internal/helpers/use-error-handler.ts
@@ -1,0 +1,35 @@
+import { useEffect, useRef } from 'react'
+
+export function useErrorHandler(
+  handleOnUnhandledError: (event: WindowEventMap['error']) => void,
+  handleOnUnhandledRejection: (
+    event: WindowEventMap['unhandledrejection']
+  ) => void
+) {
+  const stacktraceLimitRef = useRef<undefined | number>()
+
+  useEffect(() => {
+    try {
+      const limit = Error.stackTraceLimit
+      Error.stackTraceLimit = 50
+      stacktraceLimitRef.current = limit
+    } catch {}
+
+    window.addEventListener('error', handleOnUnhandledError)
+    window.addEventListener('unhandledrejection', handleOnUnhandledRejection)
+    return () => {
+      if (stacktraceLimitRef.current !== undefined) {
+        try {
+          Error.stackTraceLimit = stacktraceLimitRef.current
+        } catch {}
+        stacktraceLimitRef.current = undefined
+      }
+
+      window.removeEventListener('error', handleOnUnhandledError)
+      window.removeEventListener(
+        'unhandledrejection',
+        handleOnUnhandledRejection
+      )
+    }
+  }, [handleOnUnhandledError, handleOnUnhandledRejection])
+}
